Blog, Post, Folder, Category, User Classes
Create, edit, delete, pulldown menus, class functions, login and logout,
interface links

+++ user/interface.php
@@ -1,13 +1,26 @@
+
+<?php require_once "../Layout/header.php"; ?>
+    
 <title>Users</title>
 <h1>Users</h1>
 
+<a href="login.php" target="_self">Login</a>
+<br>
+<a href="logout.php" target="_self">Logout</a>
+<br>
 <a href="new.php" target="_self">Create New</a>
 <br>
 <a href="edit.php" target="_self">Edit User</a>
 <br>
-<a href="delete.php" target="_self">Delete User</a>
+<a href="delete.php" target="_self" onclick="return confirm('Are you sure you want to delete your account?');">Delete User</a>
 <br>
-
+<a href="../post/interface.php" target="_self">Go to Posts</a>
+<br>
+<a href="../blog/interface.php" target="_self">Go to Blogs</a>
+<br>
+<a href="../folder/interface.php" target="_self">Go to Folders</a>
+<br>
+<a href="../category/interface.php" target="_self">Go to Categories</a>
